requirements/development.txt
-r base.txt
django-debug-toolbar>=4.4
django-extensions>=3.2
pytest-django>=4.8
ruff>=0.8
